pyproject.toml
[project]
name = "cmip6-leap-feedstock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11.0"
dependencies = [
    "aiohttp>=3.12.15",
    "intake-esm>=2025.7.9",
    "requests>=2.32.4",
]

[tool.ruff.lint]
ignore = ["E712"] #ignore those shitty pandas lines in scripts/dump_bigquery_to_csv.py
